pyproject.toml
[tool.poetry]
name = "bankwidget"
version = "0.1.0"
description = ""
authors = ["Roman.Iastremskyi <iastremskyi010397@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.0"
black = "^24.4.2"
mypy = "^1.11.0"
isort = "^5.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
types-requests = "^2.32.0.20240712"

[tool.black]
line-length = 119
exclude = '''
(
  /(
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
)
'''

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = '.venv'

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
